Graphics Engine Rework Checkpoint
- Rework to support animations and transforms
- Removed scene graph state (clip rect, offsets, etc) from renderer.
- Split frame into layout, style computation, paint and composite.
- Added preload of images and exposed capinsets support.
- Fixed various bugs.

+++ packages/light-source/addon/light-source/StyleMacros.h
@@ -24,7 +24,7 @@
         return this->JSGetter(this->m_##NAME, info.Env()); \
     } \
     void set_##NAME(const Napi::CallbackInfo& info, const Napi::Value& value) { \
-        JSSetter<int64_t>(this->m_##NAME, value, false); \
+        JSSetter<uint32_t>(this->m_##NAME, value, false); \
     } \
     const StyleColorValue* NAME() const { return this->m_##NAME.get(); }
